mongoose populate & bug fix

+++ src/controllers/videoController.js
@@ -1,3 +1,4 @@
+import User from "../models/User";
 import Video from "../models/Video";
 
 export const home = async (req, res) => {
@@ -41,34 +42,49 @@ export const search = async (req, res) => {
 export const watch = async (req, res) => {
   const { id } = req.params;
 
-  const video = await Video.findById(id);
+  const video = await Video.findById(id).populate("owner");
 
   if (!video) {
     return res.render("404", { pageTitle: "Video not found." });
   }
 
-  return res.render("watch", { video });
+  return res.render("video/watch", { video });
 };
 
 export const getEdit = async (req, res) => {
   const { id } = req.params;
+  const {
+    user: { _id },
+  } = req.session;
+
   const video = await Video.findById(id);
 
   if (!video) {
     return res.status(404).render("404", { pageTitle: "Video not found." });
   }
 
-  return res.render("edit", { pageTitle: `Edit: ${video.title}`, video });
+  if (video.owner + "" !== _id + "") {
+    return res.status(403).redirect("/");
+  }
+
+  return res.render("video/edit", { pageTitle: `Edit: ${video.title}`, video });
 };
 
 export const postEdit = async (req, res) => {
   const { id } = req.params;
   const { title, description, hashtags } = req.body;
+  const {
+    user: { _id },
+  } = req.session;
 
   const video = await Video.exists({ _id: id });
 
   if (!video) {
     return res.status(404).render("404", { pageTitle: "Video not found." });
+  }
+  //Forbiding access to edit if the user is not the owner of video
+  if (video.owner + "" !== _id + "") {
+    return res.status(403).redirect("/");
   }
 
   await Video.findByIdAndUpdate(id, {
@@ -81,7 +97,7 @@ export const postEdit = async (req, res) => {
 };
 
 export const getUpload = (req, res) => {
-  return res.render("upload");
+  return res.render("video/upload");
 };
 
 export const postUpload = async (req, res) => {
@@ -94,7 +110,7 @@ export const postUpload = async (req, res) => {
   } = req;
 
   try {
-    await Video.create({
+    const newVideo = await Video.create({
       fileUrl,
       title,
       description,
@@ -102,21 +118,39 @@ export const postUpload = async (req, res) => {
       owner: _id,
     });
 
+    const user = await User.findById(_id);
+    user.videos.push(newVideo._id);
+    user.save();
+
     return res.redirect("/");
   } catch (error) {
     console.log(error);
-    return res.status(400).render("upload", { errorMessage: error._message });
+    return res
+      .status(400)
+      .render("video/upload", { errorMessage: error._message });
   }
 };
 
 export const deleteVideo = async (req, res) => {
   const { id } = req.params;
+  const {
+    user: { _id },
+  } = req.session;
 
-  const video = await Video.exists({ _id: id });
+  const video = await Video.findById(id);
 
   if (!video) {
     return res.status(404).render("404", { pageTitle: "Video not found." });
   }
+
+  if (video.owner + "" !== _id + "") {
+    return res.status(403).redirect("/");
+  }
+
+  const user = await User.findById(video.owner);
+  const newVideos = user.videos.filter((video) => video + "" !== id);
+  user.videos = newVideos;
+  user.save();
 
   await Video.findByIdAndDelete(id);
